Tag: V00-08-02 2014-01-07 Andy Salnikov - regenerate whole thing from psddldata V01-00-01 with psddl V00-13-02 - fix for Acqiris::TdcData writing
git-svn-id: https://pswww.slac.stanford.edu/svn-readonly/psdmrepo@7510 b967ad99-d558-0410-b138-e0f6c56caec7

+++ psddl_hdf2psana/trunk/include/acqiris.ddl.h
@@ -166,7 +166,7 @@ struct dataset_data {
   dataset_data(const Psana::Acqiris::TdcChannel& psanaobj);
   ~dataset_data();
 
-  uint32_t channel;
+  int32_t channel;
   uint32_t _mode_int;
   uint16_t slope;
   uint16_t mode;
@@ -258,7 +258,7 @@ struct dataset_data {
   dataset_data(const Psana::Acqiris::TdcDataV1_Item& psanaobj);
   ~dataset_data();
 
-  int32_t source;
+  uint8_t source;
   uint8_t overflow;
   uint32_t value;
